requirements.txt

torch>=2.1.1,<3.0
transformers>=4.39.0,<5.0.0
asyncio>=3.4.3,<4.0
langchain-community
langchain-experimental==0.0.55
backoff==2.2.1
toml
pypdf==4.3.1
ratelimit==2.2.1
loguru==0.7.2
pydantic==2.8.2
tenacity==8.5.0
Pillow==10.4.0
rich
psutil
sentry-sdk
python-dotenv
opencv-python-headless
PyYAML
docstring_parser==0.16
black>=23.1,<25.0
ruff>=0.0.249,<0.4.5
types-toml>=0.10.8.1
types-pytz>=2023.3,<2025.0
types-chardet>=5.0.4.6
mypy-protobuf>=3.0.0
pytest>=8.1.1
termcolor>=2.4.0
pandas>=2.2.2
fastapi>=0.110.1
networkx
swarms-memory
pre-commit
aiofiles
swarm-models
clusterops
reportlab
doc-master
